+++ app/src/main/java/com/denizd/cashmere/model/Outfit.kt
@@ -7,5 +7,5 @@ import androidx.room.PrimaryKey
 @Entity
 data class Outfit(
     val name: String,
-    @PrimaryKey(autoGenerate = true) @ColumnInfo(name = "outfit_id") val outfitId: Int
+    @PrimaryKey(autoGenerate = true) @ColumnInfo(name = "outfit_id") val outfitId: Long
 )
